requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
certifi==2019.3.9
cffi==1.12.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
decorator==4.3.2
dj-database-url==0.5.0
Django==2.1.7
django-extensions==2.1.6
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.1
djangorestframework-stubs==0.1.0
gunicorn==19.9.0
idna==2.8
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.3.4
paypalrestsdk==1.13.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.2
views==0.3
wcwidth==0.1.7
whitenoise==4.1.2
